requirements.txt
google-adk>=0.3.0
google-genai>=1.10.0
python-dotenv>=1.0.0